+++ src/parser/Occupancy.java
@@ -6,6 +6,6 @@ public interface Occupancy
 {
 	public List<String> getLabList();
 	
-	public Map<Date, Integer> getAbsoluteOccupancy(String lab);
-	public Map<Date, Float>	getRelativeOccupancy(String lab);
+	public Map<Date, Double> getAbsoluteOccupancy(String lab);
+	public Map<Date, Double>	getRelativeOccupancy(String lab);
 }
